requirements.txt
absl-py==2.1.0
accelerate==0.24.1
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncmy==0.2.9
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
deepdiff==7.0.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
email_validator==2.1.2
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
filetype==1.2.0
Flask==2.3.2
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fschat==0.2.35
fsspec==2024.6.0
fvcore==0.1.5.post20221221
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
heartrate==0.2.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
install==1.3.5
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.5
langchain-chroma==0.1.1
langchain-community==0.2.5
langchain-core==0.2.9
langchain-experimental==0.0.61
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.80
language_data==1.2.0
layoutparser==0.3.4
llama-cloud==0.0.6
llama-index==0.10.49
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.49
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.34.0
opencv-python==4.10.0.84
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.2
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.23.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.19.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
SQLAlchemy==2.0.31
srsly==2.4.8
sse-starlette==1.8.2
starlette==0.37.2
striprtf==0.0.26
svgwrite==1.4.3
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.16.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.15.2
tomli==2.0.1
# torch==2.3.1
# torchvision==0.18.1
tqdm==4.66.4
transformers==4.37.2
triton==2.3.1
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.5
unstructured-client==0.23.7
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.6
wikiextractor==0.1
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yacs==0.1.8
yarl==1.9.4
zhipuai==2.1.1.20240620.1
zipp==3.19.2
